# apps/physics_agent_service/openapi.yaml
openapi: 3.1.0
info:
  title: Physics Agent Service
  description: "The Physics Agent Service takes a USD file (uploaded directly or\
    \ referenced by S3 URI) as input and runs a multi-step VLM-based classification\
    \ pipeline: build_dataset_usd \u2192 prepare_dataset \u2192 predict \u2192 apply_physics.\
    \ It produces predictions (JSONL), an HTML report, and a simulation-ready USD\
    \ with UsdPhysics schemas applied for each prim in the scene.\n\n## Highlights\n\n\
    * End-to-end pipeline\n    * POST a USD file (multipart) or an S3 URI to start the pipeline.\n\
    \    * Two-step upload supported via POST /pipeline/upload-usd followed by POST /pipeline with session_id.\n\
    \    * Produces predictions.jsonl, an HTML classification report, and a simulation-ready\
    \ USD (scene_physics.<output extension>) with UsdPhysics RigidBodyAPI / CollisionAPI\
    \ / MassAPI / MaterialAPI authored on each predicted prim.\n* Flexible\
    \ configuration\n    * Server-side defaults control rendering modes, VLM prompts,\
    \ and prediction settings.\n    * Supports multiple rendering modes (prim_only, composition,\
    \ etc.).\n    * Configurable VLM backend and model.\n    * Selectable rendering\
    \ backend via API parameter:\n        * `remote` (default in the bundled Docker\
    \ Compose) \u2014 HTTP render service. The public compose points this at the OVRTX\
    \ sidecar, so the main service stays CPU-only.\n        * `warp` \u2014 In-process\
...
paths:
...
  /artifacts/{session_id}/predictions:
    get:
      tags:
      - artifacts
      summary: Download Predictions
      description: Download predictions JSONL file.
      operationId: download_predictions_artifacts__session_id__predictions_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /artifacts/{session_id}/report:
    get:
      tags:
      - artifacts
      summary: View Prediction Report
      description: 'View prediction HTML report in browser.


        Generates the report on-demand if it doesn''t exist yet.'
      operationId: view_prediction_report_artifacts__session_id__report_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /artifacts/{session_id}/dataset:
    get:
      tags:
      - artifacts
      summary: Download Dataset
      description: Download dataset JSONL file.
      operationId: download_dataset_artifacts__session_id__dataset_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /tune/{session_id}/artifacts/{name}:
    get:
      tags:
      - tune
      summary: Download Tune Artifact
      description: |
        Download one of the canonical tune artifacts:
        `best_params.json`, `tune_results.json`, `history.jsonl`,
        `report.md`, `tuned_physics.usd`, or the optional visual judge
        contact sheet `comparison.png`.
      operationId: download_tune_artifact_tune__session_id__artifacts__name__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Artifact bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: |
                  Raw artifact bytes. Concrete content type depends on the
                  requested artifact: `best_params.json` /
                  `tune_results.json` are JSON, `history.jsonl` is
                  JSON-lines, `report.md` is markdown,
                  `tuned_physics.usd` is a USD layer, and
                  `comparison.png` is a PNG image when visual judging ran.
                  Generated SDKs should treat the response as a binary
                  download.
        '404':
          description: Unknown artifact or session not found
...
  /refine/{session_id}/artifacts/{name}:
    get:
      tags:
      - refine
      summary: Download Refine Artifact
      description: |
        Download one of the canonical refine artifacts:
        `refine_summary.json`, `final/scenario.yaml`, `final/best_params.json`,
        `final/tune_results.json`, `final/history.jsonl`,
        `final/judge_result.json`, `final/tuned_physics.usd`,
        `final/recording.usd`, `final/report.md`, or `final/comparison.png`.
      operationId: download_refine_artifact_refine__session_id__artifacts__name__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Artifact bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: Raw artifact bytes. Concrete content type depends on the requested artifact.
        '404':
          description: Unknown artifact or session not found
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...